Implementing api rest calls.

+++ api-rest/src/main/java/com/caremao/codebreaker/api/model/WordMatchResponse.java
@@ -1,3 +1,3 @@
 package com.caremao.codebreaker.api.model;
 
-public record WordMatchResponse(String word, int matches) {}
+public record WordMatchResponse(String word, long matches) {}
